task-service/requirements.txt
flask
mysql-connector-python
requests
flask-cors